backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cloudinary==1.44.1
confusable-homoglyphs==3.3.1
Django==6.0.1
django-registration==5.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2==2.9.11
PyJWT==2.10.1
python-decouple==3.8
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
